Move opening hours to integrations
