+++ tests/test_external_services/test_hauki/test_hauki_link_generator.py
@@ -7,7 +7,7 @@ import pytest
 from django.conf import settings
 from freezegun import freeze_time
 
-from tilavarauspalvelu.utils.opening_hours.hauki_link_generator import generate_hauki_link
+from tilavarauspalvelu.integrations.opening_hours.hauki_link_generator import generate_hauki_link
 
 VALID_SIGNATURE = "87cbd7cc4f1730cb71094d2648e79045d0dde31e1025695b69f4d76c301e4f20"
 ORGANIZATION = settings.HAUKI_ORGANISATION_ID
