+++ tests/test_external_services/test_tprek/test_tprek_unit_hauki_resource_importer.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import pytest
 
+from tilavarauspalvelu.integrations.opening_hours.hauki_api_client import HaukiAPIClient
 from tilavarauspalvelu.utils.importers.tprek_unit_importer import TprekUnitHaukiResourceIdImporter
-from tilavarauspalvelu.utils.opening_hours.hauki_api_client import HaukiAPIClient
 
 from tests.factories import OriginHaukiResourceFactory, UnitFactory
 from tests.helpers import patch_method
