+++ tests/test_graphql_api/test_reservation_unit/test_create.py
@@ -6,9 +6,9 @@ import pytest
 
 from tilavarauspalvelu.enums import PaymentType, ReservationKind, WeekdayChoice
 from tilavarauspalvelu.exceptions import HaukiAPIError
+from tilavarauspalvelu.integrations.opening_hours.hauki_api_client import HaukiAPIClient
+from tilavarauspalvelu.integrations.opening_hours.hauki_api_types import HaukiAPIResource, HaukiTranslatedField
 from tilavarauspalvelu.models import ReservationUnit
-from tilavarauspalvelu.utils.opening_hours.hauki_api_client import HaukiAPIClient
-from tilavarauspalvelu.utils.opening_hours.hauki_api_types import HaukiAPIResource, HaukiTranslatedField
 
 from tests.factories import ReservationUnitPaymentTypeFactory, UnitFactory
 from tests.helpers import patch_method
