+++ tests/test_querysets/test_reservation_querysets.py
@@ -6,8 +6,8 @@ from collections import defaultdict
 import pytest
 
 from tilavarauspalvelu.enums import Language, ReservationStateChoice, ReservationTypeChoice
+from tilavarauspalvelu.integrations.opening_hours.time_span_element import TimeSpanElement
 from tilavarauspalvelu.models import AffectingTimeSpan, Reservation, ReservationUnit, ReservationUnitHierarchy
-from tilavarauspalvelu.utils.opening_hours.time_span_element import TimeSpanElement
 from utils.date_utils import DEFAULT_TIMEZONE, local_date
 
 from tests.factories import ReservationFactory, ReservationUnitFactory, ResourceFactory, SpaceFactory, UnitFactory
