+++ tilavarauspalvelu/admin/reservation_unit/admin.py
@@ -14,9 +14,9 @@ from tilavarauspalvelu.admin.application_round_time_slot.admin import Applicatio
 from tilavarauspalvelu.admin.reservation_unit_image.admin import ReservationUnitImageInline
 from tilavarauspalvelu.admin.reservation_unit_pricing.admin import ReservationUnitPricingInline
 from tilavarauspalvelu.enums import ReservationKind
+from tilavarauspalvelu.integrations.opening_hours.hauki_resource_hash_updater import HaukiResourceHashUpdater
 from tilavarauspalvelu.models import ReservationUnit
 from tilavarauspalvelu.services.csv_export import ReservationUnitExporter
-from tilavarauspalvelu.utils.opening_hours.hauki_resource_hash_updater import HaukiResourceHashUpdater
 from utils.sentry import SentryLogger
 
 from .form import ReservationUnitAdminForm
