+++ tilavarauspalvelu/integrations/opening_hours/hauki_api_types.py
