+++ tilavarauspalvelu/integrations/opening_hours/hauki_link_generator.py
