+++ tilavarauspalvelu/integrations/opening_hours/summaries.py
