+++ tilavarauspalvelu/utils/first_reservable_time/first_reservable_time_reservation_unit_helper.py
@@ -4,12 +4,12 @@ import datetime
 from typing import TYPE_CHECKING
 
 from tilavarauspalvelu.enums import ReservationStartInterval
+from tilavarauspalvelu.integrations.opening_hours.time_span_element import TimeSpanElement
+from tilavarauspalvelu.integrations.opening_hours.time_span_element_utils import merge_overlapping_time_span_elements
 from tilavarauspalvelu.utils.first_reservable_time.first_reservable_time_reservable_time_span_helper import (
     ReservableTimeSpanFirstReservableTimeHelper,
 )
 from tilavarauspalvelu.utils.first_reservable_time.utils import ReservableTimeOutput
-from tilavarauspalvelu.utils.opening_hours.time_span_element import TimeSpanElement
-from tilavarauspalvelu.utils.opening_hours.time_span_element_utils import merge_overlapping_time_span_elements
 from utils.date_utils import local_datetime, local_datetime_max, local_datetime_min, local_start_of_day
 
 if TYPE_CHECKING:
